Fix _updates thread duplicated at MQTT reconnect
When connection to the MQTT broker is lost and later reconnected, all
actions in `_on_mqtt_connect` would run again. This is correct for the
MQTT-related actions such as subscribing to the control topics, but not
for `ip.get_updates`.

The `ip150` module did not check whether the thread was already running
and just started a new one. In doing so, it overwrote the variable
referencing it, so it couldn't even be stopped later.